// cgoml-ui/src/app/features/models/messaging/views/full/MessageValidationResultFullView.ts
import { MessageValidationResultForm } from 'src/app/features/models/messaging/forms/MessageValidationResultForm';
/**
 * auto generated full veiw bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class MessageValidationResultFullView {

id: number;
form: MessageValidationResultForm;
canUpdate: boolean;
canDelete: boolean;

/* Specific Code Start */
/* Specific Code End */
}